refactor(transform): clarify oxc preeval pipeline
