+++ PLAN.tasks/TASK-08-preeval-runtime-pipeline.md
@@ -1,6 +1,6 @@
 # TASK-08: Preeval and Runtime Pipeline
 
-Status: pending
+Status: done
 
 ## Goal
 
@@ -14,17 +14,19 @@ Make `collectOxcRuntime.ts` and `oxcPreevalStage.ts` explicit pipeline facades t
 
 ## Checklist
 
-- [ ] Map current runtime collection and preeval stages.
-- [ ] Extract named pipeline steps where it improves ownership.
-- [ ] Keep strategy branching explicit and tested.
-- [ ] Ensure `execute` bypasses static adapters.
-- [ ] Ensure `hybrid` and `static` consume resolved static overlays at the right phase.
-- [ ] Run focused tests.
-- [ ] Update `PLAN.md` status and progress log.
+- [x] Map current runtime collection and preeval stages.
+- [x] Extract named pipeline steps where it improves ownership.
+- [x] Keep strategy branching explicit and tested.
+- [x] Ensure `execute` bypasses static adapters.
+- [x] Ensure `hybrid` and `static` consume resolved static overlays at the right phase.
+- [x] Run focused tests.
+- [x] Update `PLAN.md` status and progress log.
 
 ## Progress Log
 
 - 2026-05-08 14:36 EEST: Task created by scaffold.
+- 2026-05-08 16:14 EEST: Started Task 08. Current files are under the size limit; scope is ownership clarity for runtime normalization/source maps and preeval strategy/code-preparation steps.
+- 2026-05-08 16:24 EEST: Split runtime normalization/source maps and preeval processor collection, strategy overlay, code preparation, and `__wywPreval` export into explicit modules. Focused tests, transform typecheck, lint, and size guard pass.
 
 ## Context Recovery Notes
