+++ packages/transform/src/utils/oxcPreevalStage/evalStrategy.ts
@@ -1,0 +1,7 @@
+import type { OxcPreevalOptions } from './types';
+
+export const getEvalStrategy = (options: OxcPreevalOptions) =>
+  options.eval?.strategy ?? 'hybrid';
+
+export const usesStaticEvaluation = (options: OxcPreevalOptions): boolean =>
+  getEvalStrategy(options) !== 'execute';
